update: Comment, LostFoundBoard 컨트롤러명 변경
